requirements.txt
# Core scraping
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.1.0

# Dynamic scraping engines
selenium>=4.15.0
playwright>=1.40.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP & networking
urllib3>=2.1.0
fake-useragent>=1.4.0


# Database (PostgreSQL)
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
alembic>=1.13.0

# Resilience & retry
tenacity>=8.2.0

# Logging & monitoring
colorlog>=6.8.0
rich>=13.7.0


# Configuration
python-dotenv>=1.0.0# Caching
diskcache>=5.6.0# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
responses>=0.24.0
